4.6.8
- Added Tugboat.ReuseAddress.
- Fixed nested NetworkObjects not spawning when pooling.
- Improved added null-check to SyncTypes with collections when user initializes with their own collection.
- Improved Object/CollectionCaches moved null checks from StoreAndDefault to Store.
- Improved ResettableCollectionCaches resets call ResettableObjectCaches on each entry rather than handling locally.
- Changed LevelLoggingConfiguration.EnableLogging to IsEnabled.
- Added LevelLoggingConfiguration.AddLocalTick.
- Added LevelLoggingConfiguration.AddTimestamps.
- Added LevelLoggingConfiguration.EnableTimestampsInEditor.
- Improved SyncTypeSettings are now serializable.

+++ Assets/FishNet/Runtime/Managing/Statistic/NetworkTrafficStatistics.cs
@@ -167,7 +167,7 @@ namespace FishNet.Managing.Statistic
         /// <summary>
         /// Formats passed in bytes value to the largest possible data type with 2 decimals.
         /// </summary>
-        public static string FormatBytesToLargest(ulong bytes)
+        public static string FormatBytesToLargest(float bytes)
         {
             int decimalPlaces = 2;
             if (bytes == 0)
